+++ firmware/lib/Sensors/TempHumiSensor.h
@@ -22,7 +22,7 @@ class TempSHTSensor final : public SharingSensor<SharedSHTSensor>
 public:
     using SharingSensor<SharedSensor>::SharingSensor;
     SensorValue getMeasurement();
-    uint32_t getTypeTag() const { return TEMP_SHT_KEY; }
+    uint8_t getTypeTag() const { return TEMP_SHT_KEY; }
 };
 
 class HumiSHTSensor final : public SharingSensor<SharedSHTSensor>
@@ -30,6 +30,6 @@ class HumiSHTSensor final : public SharingSensor<SharedSHTSensor>
 public:
     using SharingSensor<SharedSensor>::SharingSensor;
     SensorValue getMeasurement();
-    uint32_t getTypeTag() const { return HUMI_SHT_KEY; }
+    uint8_t getTypeTag() const { return HUMI_SHT_KEY; }
 };
 #endif
